seperated out link handler so it can be explicitly called if preventing propogation
