+++ js/templates/components/moderatorCard.html
@@ -5,7 +5,7 @@
    const style = ob.verified ? 'verified clrBrAlert2 clrBAlert2Grad' : '';
 %>
 
-<div class="moderatorCardInner js-moderatorCard clrP <%= isDisabled %> <%= style %>">
+<div class="moderatorCardInner clrP <%= isDisabled %> <%= style %>">
   <div class="flexRow gutterH moderatorCardContent">
     <% if (ob.radioStyle) { %>
       <div class="flexNoShrink">
